change paymentTerm to int

+++ src/main/java/com/velotrade/sdk/entity/Auction.java
@@ -17,7 +17,7 @@ public class Auction {
     private String currency;
     private float amount;
     private float expectedAmount;
-    private float paymentTerms;
+    private int paymentTerms;
     private Date dueDate;
     private Date expectedPaymentDate;
     private String description;
@@ -120,11 +120,11 @@ public class Auction {
         this.expectedAmount = expectedAmount;
     }
 
-    public float getPaymentTerms() {
+    public int getPaymentTerms() {
         return paymentTerms;
     }
 
-    public void setPaymentTerms(float paymentTerms) {
+    public void setPaymentTerms(int paymentTerms) {
         this.paymentTerms = paymentTerms;
     }
